backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
Pillow
reportlab
pydicom
numpy
python-dotenv
requests
